Use `litrs` for string literal parsing
While this adds another dependency and a bit of compile time, using
a dedicated crate for literal parsing has many advantages. For one,
litrs is a lot better tested and can focus on this single task. But we
also don't have duplicate code that has to be maintained as part of
`bunt`. Finally. the bunt code surely was not as robust as the litrs
code and failed for a couple inputs.