Add `[set_]std*_color_choice` functions to avoid hardcoding `Auto`
The previous advice was to use `write[ln]` in those cases where the
default is not sufficient, but I think a majority of use cases can
be solved by just providing this global config. It's the case in a
project I'm using bunt in.

This of course adds a bit of overhead to every `println` call, but I
am fairly sure it's far less than what a `print` costs anyway.